fix: reset polling interval to the user default setting when a request does not fail
