+++ src/store/network/utils.ts
@@ -5,11 +5,62 @@
  */
 
 import { forEach } from 'lodash';
-import { SoapContext } from '../../../types';
+import { AccountSettings, SoapContext, SoapResponse } from '../../../types';
 import { folderWorker, tagWorker } from '../../workers';
 import { useFolderStore } from '../folder';
 import { useTagStore } from '../tags';
 import { useNetworkStore } from './store';
+import { useAccountStore } from '../account';
+
+/**
+ * Polling interval to use if the long polling delay
+ * is not allowed for the user
+ */
+const POLLING_NOWAIT_INTERVAL = 10_000;
+
+/**
+ * Polling interval to use if a previous request failed
+ * with a 500 error
+ */
+const POLLING_RETRY_INTERVAL = 60_000;
+
+export const parsePollingInterval = (settings: AccountSettings): number => {
+	const pollingPref = (settings.prefs?.zimbraPrefMailPollingInterval ?? '') as string;
+	if (pollingPref === '500') {
+		return 500;
+	}
+	const pollingValue = parseInt(pollingPref, 10);
+	if (Number.isNaN(pollingValue)) {
+		return 30000;
+	}
+	if (pollingPref.includes('m')) {
+		return pollingValue * 60 * 1000;
+	}
+	return pollingValue * 1000;
+};
+
+/**
+ * Return the polling interval for the next NoOp request.
+ * The interval length depends on the user settings, but it can be
+ * overridden by the server response/errors
+ * @param res
+ */
+export const getPollingInterval = <R>(res: SoapResponse<R>): number => {
+	const { pollingInterval } = useNetworkStore.getState();
+	const { settings } = useAccountStore.getState();
+	const waitDisallowed = (res?.Body as { waitDisallowed?: number })?.waitDisallowed;
+	const fault = res?.Body?.Fault;
+	if (fault) {
+		return POLLING_RETRY_INTERVAL;
+	}
+	if (waitDisallowed) {
+		return POLLING_NOWAIT_INTERVAL;
+	}
+	if (!fault) {
+		return parsePollingInterval(settings);
+	}
+	return pollingInterval;
+};
 
 export const handleSync = ({ refresh, notify }: SoapContext): Promise<void> =>
 	new Promise((r) => {
